Daemon implementation finished

+++ python/modules/stop_model.py
@@ -6,11 +6,18 @@ from utils.vars import RUNTIME_DAEMON_DATA
 from utils.llama_cleanup import *
 
 def stop_model(model):
-    info(f"Stopping model {model}...")
-    pid = RUNTIME_DAEMON_DATA[model]["pid"]
-    if pid:
-        os.kill(pid, signal.SIGTERM)
-        success(f"Server for model {model} succesfully killed. PID {pid}")
-        llama_cleanup(model)
+    if RUNTIME_DAEMON_DATA["server"]["running"]:
+        info(f"Stopping model {model}...")
+        models_list = RUNTIME_DAEMON_DATA["server"]["models"]
+        if model in models_list:
+            pid = RUNTIME_DAEMON_DATA[model]["pid"]
+            if pid:
+                os.kill(pid, signal.SIGTERM)
+                success(f"Server for model {model} succesfully killed. PID {pid}")
+                llama_cleanup(model)
+                return True
+        else:
+            error(f"Unable to find model with the name {model}...")
+            return False
     else:
-        error(f"Unable to find model with the name {model}...")
+        error("Gekoku daemon is not running")
